ballerina-tools 0.964.0 (new formula)
Installs the complete Ballerina toolset, including the composer.
Replaces (and aliases) Formula/ballerina.rb
